% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.r
\name{compute_impulse_responses}
\alias{compute_impulse_responses}
\title{Function to draw one single path for the impulse-response functions.}
\usage{
compute_impulse_responses(Alpha, Sigma, nolags, nhor)
}
\arguments{
\item{Alpha}{the estimate of the VAR-coefficients}

\item{Sigma}{the estimate of the Variance-covariance matrix}

\item{nolags}{Number of lags in the model}

\item{nhor}{the horizon of the impulse-response functions}
}
\description{
Function to draw one single path for the impulse-response functions.
}
