% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast.ltvar}
\alias{forecast.ltvar}
\title{forecasting using an ltvar model}
\usage{
\method{forecast}{ltvar}(obj, n.ahead, ...)
}
\arguments{
\item{obj}{a fitted ltvar model}

\item{n.ahead}{forecast horizon}

\item{...}{currently not used}
}
\description{
forecasting using an ltvar model
}
