% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsamplt.r
\name{fSampLT}
\alias{fSampLT}
\title{sample the latent thresholds}
\usage{
fSampLT(my, amX, mp, amG2inv, vd, vdc)
}
\arguments{
\item{my}{matrix with the data (lhs)}

\item{amX}{matrix with rhs of data}

\item{mp}{ar-coefficient}

\item{amG2inv}{volatility}

\item{vd}{vector with current treshold}

\item{vdc}{vector with candidate thresholds}
}
\description{
sample the latent thresholds
}
