% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fSampLTBv}
\alias{fSampLTBv}
\title{samples}
\usage{
fSampLTBv(my, amX, mPhi, vmu, mSig, amG2inv, vd, mbo, stabletest,
  Intercept, nl)
}
\arguments{
\item{my}{LHS of data}

\item{mPhi}{transition matrix for VAR-Parameters}

\item{mx}{RHS of data}
}
\description{
samples
}
