% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsamspg.r
\name{fSampSPG}
\alias{fSampSPG}
\title{samples sigma, phi and gamma for stochastic volatility sampler}
\usage{
fSampSPG(my, mh, mPhi, vgam, dnu, dV0, da0, db0, dg0, dG0)
}
\arguments{
\item{my}{current draw of stochastic volatility}

\item{mh}{previous draw of stochastic volatility}

\item{mPhi}{Phi-matrix from previous draw}

\item{vgam}{gamma draw from previous round}

\item{dnu, dV0}{prior parameters on sigma, inverse gamma distribution}

\item{da0, db0}{prior parameters on phi, shape parameters for beta distribution}

\item{dg0, dG0}{prior parameters on gamma, inverse gamma distribution}
}
\description{
samples sigma, phi and gamma for stochastic volatility sampler
}
