% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fSampSPM}
\alias{fSampSPM}
\title{samples the parameters for the transition of VAR-parameters}
\usage{
fSampSPM(mp, mSig, mPhi, vmu, dnu, dV0, da0, db0, dm0, ds0, vd, dk0)
}
\arguments{
\item{mp}{VAR-parameters}

\item{mSig}{standard deviation of the paramters}

\item{mPhi}{matrix with the transition parameters}

\item{vmu}{vector with intercepts}

\item{dnu}{prior on the volatility of parameters, scale of inverse gamma distribution}

\item{dV0}{prior on the volatility of parameters, shape of inverse gamma distribution}

\item{da0}{prior on Phi, shape of beta distribution (alpha)}

\item{db0}{prior on Phi, shpae of beta distribution (beta)}

\item{dm0}{Prior on Intercept, mean of normal distribution}

\item{ds0}{Prior on Intercept, variance of normal distribution}

\item{vd}{vector with the thresholds}

\item{dk0}{latent threshold prior}
}
\description{
samples the parameters for the transition of VAR-parameters
}
