% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsampsv.r
\name{fSVSamp}
\alias{fSVSamp}
\title{implements multi-move sampler for SV model}
\usage{
fSVSamp(vy, vh, dphi, dsig2, dh00, dsig02, nK = NULL)
}
\arguments{
\item{vy}{response (ns*1 vector)}

\item{vh}{current point of h (ns*1 vector)}

\item{dphi}{AR(1) coefficient for transition of volatility}

\item{dsig2, dsig02}{parameters (scalar)}

\item{dh00}{mean of cofficients}

\item{nK}{number of blocks for multi-move sampler}
}
\description{
implements multi-move sampler for SV model
}
