% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{jpdata}
\alias{jpdata}
\title{Japanese Macroeconomic data}
\format{A ts objet with japanese macroeconomic from 1977 Q1 to 2007 Q4
\describe{
\item{Inflation}{Inflation}
\item{GDP gap}{GDP gap}
\item{Interest Rate Gap}{Interest Rate Gap}
}}
\usage{
jpdata
}
\description{
Japanese Macroeconomic data
}
\keyword{datasets}
