% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltvar.r
\name{ltvar}
\alias{ltvar}
\title{mcmc computation of latent-threshold models}
\usage{
ltvar(y, p = 2, Intercept = TRUE, nreps = 100, burnin = 10,
  dvb0 = 20, dVb0 = 0.002, dva0 = 2, dVa0 = 0.002, dvh0 = 2,
  dVh0 = 0.002, dm0 = 0, ds0 = 1, da0 = 1, db0 = 1, dg0 = 4,
  dG0 = 0.1, nKnots = NULL)
}
\arguments{
\item{y}{A TxK matrix with the data}

\item{p}{number of lags}

\item{Intercept}{Logical flag whether the model contains an intercept}

\item{nreps}{number of total mcmc draws}

\item{burnin}{number of burn-in draws}

\item{nKnots}{number of knots in the stochastic volatility sampler}
}
\description{
mcmc computation of latent-threshold models
}
