% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fSampLTBv}
\alias{fSampLTBv}
\title{samples the parameter
This function samples the VARS-parameter of a latent-threshold VAR using a single move algorithm.}
\usage{
fSampLTBv(my, amX, mPhi, vmu, mSig, amG2inv, vd, mbo, stabletest,
  Intercept, nl)
}
\arguments{
\item{my}{LHS of data}

\item{amX}{RHS of data}

\item{mPhi}{transition matrix for VAR-Parameters}

\item{vmu}{mean of transition matrix}

\item{mSig}{variance-covariance matrix for standard deviation of parameters}

\item{amG2inv}{a KxKxT array with the stochastic volatility}

\item{vd}{latent threshold}

\item{mbo}{old betas}

\item{stabletest}{Test if candidate draw is stable (currently not implemented)}

\item{Intercept}{whether the model contains an intercept or not.}

\item{nl}{number of lags}
}
\description{
samples the parameter
This function samples the VARS-parameter of a latent-threshold VAR using a single move algorithm.
}
