% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.r
\name{impulse_response.ltvar}
\alias{impulse_response.ltvar}
\title{impulse-responses for latent threshold VARs}
\usage{
\method{impulse_response}{ltvar}(obj, n.ahead)
}
\arguments{
\item{obj}{A fitted ltvar model}

\item{n.ahead}{Integer specifying the steps}
}
\description{
impulse-responses for latent threshold VARs
}
