% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltvar.r
\name{ltvar}
\alias{ltvar}
\title{MCMC computation of latent-threshold models.}
\usage{
ltvar(y, p = 2, Intercept = TRUE, nreps = 100, burnin = 10,
  dvb0 = 20, dVb0 = 0.002, dva0 = 2, dVa0 = 0.002, dvh0 = 2,
  dVh0 = 0.002, dm0 = 0, ds0 = 1, da0 = 1, db0 = 1, dg0 = 4,
  dG0 = 0.1, dk0 = 3, nKnots = NULL)
}
\arguments{
\item{y}{A TxK matrix with the data}

\item{p}{number of lags}

\item{Intercept}{Logical flag whether the model contains an intercept}

\item{nreps}{number of total mcmc draws}

\item{burnin}{number of burn-in draws}

\item{dvb0, dVb0}{prior on volatility of betas}

\item{dva0, dVa0}{prior on volatiltiy of covariances}

\item{dvh0, dVh0}{prior on volatility of variances}

\item{dm0, ds0}{prior on intercept}

\item{da0}{prior on phi for covariances}

\item{db0}{prior on phi for betas}

\item{dg0, dG0}{prior on parameters for stochastic volatility}

\item{dk0}{latent threshold prior}

\item{nKnots}{number of blocks in the stochastic volatility sampler}
}
\description{
Estimate a latent-threshold VAR model using a single-move Gibbs sampler as in Nakajima and West (2013)
}
\references{
Nakajima, J. and M. West (2013) Bayesian Analysis of Latent Threshold Dynamic Models; Journal of Business & Economic Statistics 31 (2), 151-164
}
